+++ source/extensions/omni.flux.stage_manager.core/docs/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.2.0]
+### Changed
+- Modified the schema resolving logic to resolve children plugins
+
 ## [1.1.2]
 ### Changed
 - Use updated setup function for interaction plugins