+++ source/extensions/omni.flux.stage_manager.plugin.context.usd/docs/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.3.0]
+### Added
+- Added listeners to the USD context base
+
 ## [1.2.0]
 ### Changed
 - Override updated `get_items` function instead of `setup`