+++ source/extensions/omni.flux.utils.widget/docs/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.18.0]
+### Changed
+- Added `subscribe_selection_changed` function to the `TreeWidget`
+
 ## [1.17.1]
 ### Changed
 - Improved branch building function for `TreeDelegateBase`